+++ Kernel/System/Ticket/Event/NotificationEvent.pm
@@ -34,6 +34,7 @@ our @ObjectDependencies = (
     'Kernel::System::Ticket',
     'Kernel::System::Ticket::Article',
     'Kernel::System::User',
+    'Kernel::System::Valid',
 );
 
 sub new {
@@ -252,6 +253,8 @@ sub Run {
                 next TRANSPORT;
             }
 
+            my @ValidIDs = $Kernel::OM->Get('Kernel::System::Valid')->ValidIDsGet();
+
             # check if transport is usable
             next TRANSPORT if !$TransportObject->IsUsable();
 
@@ -299,6 +302,17 @@ sub Run {
 
                     # No UserID means it's not a mapped customer.
                     next BUNDLE if !$Bundle->{Recipient}->{UserID};
+                }
+
+                # Check if customer user is invalid.
+                if (
+                    $Bundle->{Recipient}->{Source}
+                    && $Bundle->{Recipient}->{Source} eq 'CustomerUser'
+                    && $Bundle->{Recipient}->{ValidID}
+                    && !grep { $Bundle->{Recipient}->{ValidID} eq $_ } @ValidIDs
+                    )
+                {
+                    next BUNDLE;
                 }
 
                 my $Success = $Self->_SendRecipientNotification(